MGDSTRM-6057:  Added new section on updating Service Registry artifacts. Minor edits of the Downloading section. Updated some command output based on output seen locally. (Upstream PR 386 in app-service-guides)
MGDSTRM-6925: Added text about existing service accounts. Corrected an attribute name, and replaced other occurrences of the product name with attributes. Corrected some web console references. (Upstream PR 387 in app-service-guides)

MGDSTRM-7021: Added an attribute for the expanded version of the rhoas abbreviation, and included this new attribute on the first occurrence of the rhoas text. (Upstream PR 388 in app-service-guides)